+++ CHANGELOG.md
@@ -16,6 +16,7 @@ v0.6.0 (in development)
 - Added `GHRepo::as_str()` method
 - **Breaking**: The `GHRepo::is_valid_owner()` and `GHRepo::is_valid_name()`
   methods are now regular functions
+- Added `is_valid_repository()` function
 
 v0.5.0 (2023-04-27)
 -------------------